:recycle: Genericise and order seed pattern.
