+++ api/src/db/seeds/020_locations-seeds.ts
@@ -1,11 +1,13 @@
-import path from "path"
 import fs from "fs"
+
+import { Knex } from "knex"
 import Papa from "papaparse"
 
+import { APP_ROOT } from "@/config"
 import { Location } from "@/models"
 
-export async function locationsSeeds() {
-  const fileName = path.resolve(__dirname, "./locations-seeds.csv")
+export async function seed(knex: Knex): Promise<void> {
+  const fileName = `${APP_ROOT}/db/data/locations.csv`
   const fileContent = fs.readFileSync(fileName, "utf8")
   const { data } = Papa.parse<{ province: string; city: string }>(fileContent, {
     header: true,
@@ -15,5 +17,3 @@ export async function locationsSeeds() {
   await Location.destroy({ where: {} })
   await Location.bulkCreate(data)
 }
-
-export default locationsSeeds
