+++ api/src/db/seeds/030_travel-authorization-pre-approvals-seeds.ts
@@ -1,0 +1,5 @@
+import { Knex } from "knex"
+
+export async function seed(knex: Knex): Promise<void> {
+  // TODO
+}
